requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg==3.1.8
psycopg2-binary==2.9.5

